+++ internal/provider/resource_tfe_workspace_settings_test.go
@@ -6,15 +6,17 @@ package provider
 import (
 	"errors"
 	"fmt"
+	"math/rand"
 	"regexp"
 	"testing"
+	"time"
 
 	"github.com/hashicorp/go-tfe"
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
 	"github.com/hashicorp/terraform-plugin-testing/terraform"
 )
 
-func TestAccTFEWorkspaceSettings(t *testing.T) {
+func TestAccTFEWorkspaceSettings_basic(t *testing.T) {
 	tfeClient, err := getClientUsingEnv()
 	if err != nil {
 		t.Fatal(err)
@@ -48,6 +50,8 @@ func TestAccTFEWorkspaceSettings(t *testing.T) {
 						"tfe_workspace_settings.foobar", "overwrites.0.agent_pool", "true"),
 					resource.TestCheckResourceAttr(
 						"tfe_workspace_settings.foobar", "overwrites.#", "1"),
+					resource.TestCheckResourceAttr(
+						"tfe_workspace_settings.foobar", "auto_apply", "false"),
 				),
 			},
 			// Change to agent pool
@@ -94,7 +98,7 @@ func TestAccTFEWorkspaceSettings(t *testing.T) {
 	})
 }
 
-func TestAccTFEWorkspaceWithSettings(t *testing.T) {
+func TestAccTFEWorkspaceSettings_stateSharing(t *testing.T) {
 	tfeClient, err := getClientUsingEnv()
 	if err != nil {
 		t.Fatal(err)
@@ -116,6 +120,89 @@ func TestAccTFEWorkspaceWithSettings(t *testing.T) {
 					resource.TestCheckResourceAttrSet(
 						"tfe_workspace_settings.foobar", "workspace_id",
 					),
+				),
+			},
+		},
+	})
+}
+
+func TestAccTFEWorkspaceSettings_overlappingBooleans(t *testing.T) {
+	tfeClient, err := getClientUsingEnv()
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	org, cleanupOrg := createBusinessOrganization(t, tfeClient)
+	t.Cleanup(cleanupOrg)
+
+	resource.Test(t, resource.TestCase{
+		PreCheck:                 func() { testAccPreCheck(t) },
+		ProtoV5ProviderFactories: testAccMuxedProviders,
+		Steps: []resource.TestStep{
+			// Start with local execution
+			{
+				Config: testAccTFEWorkspaceSettingsOverlappingBooleans(org.Name),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttrSet(
+						"tfe_workspace.name", "id"),
+					resource.TestCheckResourceAttr(
+						"tfe_workspace.name", "auto_apply", "true"),
+					resource.TestCheckResourceAttr(
+						"tfe_workspace.name", "assessments_enabled", "false"),
+					resource.TestCheckResourceAttr(
+						"tfe_workspace_settings.self", "auto_apply", "true"),
+					resource.TestCheckResourceAttr(
+						"tfe_workspace_settings.self", "assessments_enabled", "false"),
+				),
+			},
+		},
+	})
+}
+
+func TestAccTFEWorkspaceSettings_basicOptions(t *testing.T) {
+	tfeClient, err := getClientUsingEnv()
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	org, cleanupOrg := createBusinessOrganization(t, tfeClient)
+	t.Cleanup(cleanupOrg)
+
+	resource.Test(t, resource.TestCase{
+		PreCheck:                 func() { testAccPreCheck(t) },
+		ProtoV5ProviderFactories: testAccMuxedProviders,
+		Steps: []resource.TestStep{
+			// Start with local execution
+			{
+				Config: testAccTFEWorkspaceSettings_options(org.Name, "initial", true),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttrSet(
+						"tfe_workspace_settings.foobar", "id"),
+					resource.TestCheckResourceAttrSet(
+						"tfe_workspace_settings.foobar", "workspace_id",
+					),
+					resource.TestCheckResourceAttr(
+						"tfe_workspace_settings.foobar", "description", "initial"),
+					resource.TestCheckResourceAttr(
+						"tfe_workspace_settings.foobar", "auto_apply", "true"),
+					resource.TestCheckResourceAttr(
+						"tfe_workspace_settings.foobar", "assessments_enabled", "true"),
+				),
+			},
+			{
+				Config: testAccTFEWorkspaceSettings_options(org.Name, "updated", false),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttrSet(
+						"tfe_workspace_settings.foobar", "id"),
+					resource.TestCheckResourceAttrSet(
+						"tfe_workspace_settings.foobar", "workspace_id",
+					),
+					resource.TestCheckResourceAttr(
+						"tfe_workspace_settings.foobar", "description", "updated"),
+					resource.TestCheckResourceAttr(
+						"tfe_workspace_settings.foobar", "auto_apply", "false"),
+					resource.TestCheckResourceAttr(
+						"tfe_workspace_settings.foobar", "assessments_enabled", "false"),
 				),
 			},
 		},
@@ -178,7 +265,7 @@ func TestAccTFEWorkspaceSettingsRemoteState(t *testing.T) {
 	})
 }
 
-func TestAccTFEWorkspaceSettingsImport(t *testing.T) {
+func TestAccTFEWorkspaceSettings_import(t *testing.T) {
 	tfeClient, err := getClientUsingEnv()
 	if err != nil {
 		t.Fatal(err)
@@ -213,7 +300,7 @@ func TestAccTFEWorkspaceSettingsImport(t *testing.T) {
 	})
 }
 
-func TestAccTFEWorkspaceSettingsImport_ByName(t *testing.T) {
+func TestAccTFEWorkspaceSettings_importByName(t *testing.T) {
 	tfeClient, err := getClientUsingEnv()
 	if err != nil {
 		t.Fatal(err)
@@ -249,6 +336,62 @@ func TestAccTFEWorkspaceSettingsImport_ByName(t *testing.T) {
 	})
 }
 
+func TestAccTFEWorkspaceSettings_basicTags(t *testing.T) {
+	rInt := rand.New(rand.NewSource(time.Now().UnixNano())).Int()
+
+	resource.Test(t, resource.TestCase{
+		PreCheck:                 func() { testAccPreCheck(t) },
+		ProtoV5ProviderFactories: testAccMuxedProviders,
+		Steps: []resource.TestStep{
+			// Start with local execution
+			{
+				Config: testAccTFEWorkspaceSettings_basicTagOne(rInt),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr(
+						"tfe_workspace_settings.test", "tags.%", "1"),
+					resource.TestCheckResourceAttr(
+						"tfe_workspace_settings.test", "tags.keyA", "valueA"),
+					resource.TestCheckResourceAttr(
+						"tfe_workspace_settings.test", "effective_tags.%", "2"),
+				),
+			},
+			{
+				Config: testAccTFEWorkspaceSettings_basicTagTwo(rInt),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr(
+						"tfe_workspace_settings.test", "tags.%", "2"),
+					resource.TestCheckResourceAttr(
+						"tfe_workspace_settings.test", "tags.keyA", "valueA"),
+					resource.TestCheckResourceAttr(
+						"tfe_workspace_settings.test", "tags.keyB", "valueB"),
+					resource.TestCheckResourceAttr(
+						"tfe_workspace_settings.test", "effective_tags.%", "3"),
+				),
+			},
+			{
+				Config: testAccTFEWorkspaceSettings_basicTagOne(rInt),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr(
+						"tfe_workspace_settings.test", "tags.%", "1"),
+					resource.TestCheckResourceAttr(
+						"tfe_workspace_settings.test", "tags.keyA", "valueA"),
+					resource.TestCheckResourceAttr(
+						"tfe_workspace_settings.test", "effective_tags.%", "2"),
+				),
+			},
+			{
+				Config: testAccTFEWorkspaceSettings_basicTagZero(rInt),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr(
+						"tfe_workspace_settings.test", "tags.%", "0"),
+					resource.TestCheckResourceAttr(
+						"tfe_workspace_settings.test", "effective_tags.%", "1"),
+				),
+			},
+		},
+	})
+}
+
 func testAccCheckTFEWorkspaceSettingsDestroy(s *terraform.State) error {
 	tfeClient, err := getClientUsingEnv()
 	if err != nil {
@@ -279,6 +422,20 @@ func testAccCheckTFEWorkspaceSettingsDestroy(s *terraform.State) error {
 	}
 
 	return nil
+}
+
+func testAccTFEWorkspaceSettingsOverlappingBooleans(orgName string) string {
+	return fmt.Sprintf(`
+resource "tfe_workspace_settings" "self" {
+  workspace_id        = tfe_workspace.name.id
+}
+
+resource "tfe_workspace" "name" {
+	organization = "%s"
+	name="permanent_auto_apply_drift"
+	auto_apply = true
+}
+`, orgName)
 }
 
 func testAccTFEWorkspaceSettingsUnknownIDRemoteState(orgName string) string {
@@ -365,4 +522,75 @@ resource "tfe_workspace_settings" "foobar" {
 	workspace_id   = "%s"
 }
 `, orgName, workspaceID)
+}
+
+func testAccTFEWorkspaceSettings_options(orgName string, description string, boolOptions bool) string {
+	return fmt.Sprintf(`
+resource "tfe_workspace" "foobar1" {
+	name = "foobar1"
+	organization = "%s"
+}
+
+resource "tfe_workspace_settings" "foobar" {
+	workspace_id              = tfe_workspace.foobar1.id
+	description               = "%s"
+	auto_apply                = %t
+	assessments_enabled       = %t
+}
+`, orgName, description, boolOptions, boolOptions)
+}
+
+func tagBase(rInt int) string {
+	return fmt.Sprintf(`
+resource "tfe_organization" "test" {
+  name  = "tst-tfeprovider-%d"
+  email = "admin@company.com"
+}
+
+resource "tfe_project" "test" {
+  organization = tfe_organization.test.name
+  name = "tfe-provider-test-%d"
+	tags = {
+	  projectTag = "valueA"
+	}
+}
+
+resource "tfe_workspace" "test" {
+	name         = "tfe-provider-test-workspace-%d"
+	organization = tfe_organization.test.name
+    project_id   = tfe_project.test.id 
+}
+`, rInt, rInt, rInt)
+}
+
+func testAccTFEWorkspaceSettings_basicTagOne(rInt int) string {
+	return tagBase(rInt) + `
+resource "tfe_workspace_settings" "test" {
+	workspace_id = tfe_workspace.test.id
+	tags = {
+	  keyA = "valueA"
+	}
+}
+`
+}
+
+func testAccTFEWorkspaceSettings_basicTagTwo(rInt int) string {
+	return tagBase(rInt) + `
+resource "tfe_workspace_settings" "test" {
+	workspace_id = tfe_workspace.test.id
+	tags = {
+	  keyA = "valueA"
+	  keyB = "valueB"	
+	}
+}
+`
+}
+
+func testAccTFEWorkspaceSettings_basicTagZero(rInt int) string {
+	return tagBase(rInt) + `
+resource "tfe_workspace_settings" "test" {
+	workspace_id = tfe_workspace.test.id
+	tags = {}
+}
+`
 }